Move filter function to viewModel
